# server/openapi.yaml
openapi: 3.0.3
info:
  title: Velo API
  version: 4.1.0
  description: |
    Project Tool for AI.

    ## Field Naming Convention
    **All request bodies use camelCase** (e.g. `ministryCode`, `fiscalYear`, `fundingSource`).
    **All responses use snake_case** DB column names (e.g. `project_name`, `ministry_code`).
    This is intentional — the controller maps camelCase input to snake_case DB fields.

    ## Authentication
    Two methods:
    - **Cookie SSO**: Sign in via `GET /auth/google`, then include `X-CSRF-Token` header on writes.
    - **API Key**: Include `X-API-Key: velo_xxx` or `Authorization: Bearer velo_xxx` header. No CSRF needed.

    API key actions are attributed to the key owner and flagged as `api_key` in the audit log.

    ## Role-Based Access Control
    Four roles with multi-role support:
    - **user**: Read-only viewer — can see projects, dashboards, and data but cannot modify anything
    - **runner**: Can DO work — velocity game, SharePoint, Git operations, run audits, edit OPEN projects (no active members), and (per project) clone/claim
    - **project_lead**: Can SET UP work — create/edit/delete projects, modules, budgets, leads, links, Applications & Contracts, persons, canvas + everything runner can do
    - **admin**: System administration — user management, settings, PAT management, project clone-policy + everything project_lead can do

    Users can hold multiple roles simultaneously. Admin role implicitly grants all permissions.
...

...

...

...

...

...

...
servers:
...
security:
  - apiKeyHeader: []
  - apiKeyBearer: []
  - cookieAuth: []
components:
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: access_token }
    apiKeyHeader: { type: apiKey, in: header, name: X-API-Key }
    apiKeyBearer: { type: http, scheme: bearer }
...
paths:
...
  /projects/{id}/audit:
    get:
      tags: [Audit]
      summary: Paginated audit log for project and all sub-resources (immutable — no DELETE endpoint)
      security: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, default: 30, maximum: 100 } }
      responses:
        200:
          description: Paginated audit entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: object
                    properties:
                      data:
                        type: array
                        items:
                          type: object
                          properties:
                            pk_audit_log: { type: string, format: uuid }
                            audit_action: { type: string, enum: [INSERT, UPDATE, DELETE] }
                            audit_table_name: { type: string }
                            audit_record_id: { type: string, format: uuid }
                            audit_new_data: { type: object, nullable: true, description: "Includes _authSource (session|api_key) and _apiKeyId" }
                            audit_old_data: { type: object, nullable: true }
                            user_display_name: { type: string }
                            created_at: { type: string, format: date-time }
                      pagination: { $ref: '#/components/schemas/PaginatedResponse/properties/pagination' }

  # ═══ Ministries ═══
...
  /projects/{id}/audits:
    get:
      tags: [Audits]
      summary: List project audits
      security: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, default: 20 } }
        - name: source
          in: query
          schema: { type: string, enum: [git, jira, confluence, sharepoint, web, manual, ai_analysis] }
      responses:
        '200':
          description: Paginated list of audits
    post:
      tags: [Audits]
      summary: Run a project audit
      description: |
        Triggers an audit from the specified source. For git audits, extracts repository analytics.
        Results are stored in the audit table and can be further analyzed with LLM.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [source, sourceUrl]
              properties:
                source: { type: string, enum: [git, jira, confluence, sharepoint, web, manual] }
                sourceUrl: { type: string, format: uri }
                moduleId: { type: string, format: uuid, description: Optional module to scope the audit to }
            examples:
              git_audit: { summary: GitHub audit, value: { source: git, sourceUrl: "https://github.com/example-org/my-app" } }
              jira_audit: { summary: Jira audit, value: { source: jira, sourceUrl: "https://example.atlassian.net/browse/PROJ-123" } }
      responses:
        '201':
          description: Audit created with results
  /projects/{id}/audits/{auditId}:
    get:
      tags: [Audits]
      summary: Get audit detail
      security: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: auditId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Full audit data including AI analysis
    delete:
      tags: [Audits]
      summary: Delete own audit
      description: Soft-deletes an audit record. Users can only delete audits they created.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: auditId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204': { description: Audit deleted }
        '403': { description: Cannot delete audits created by other users }
...
  /projects/{id}/deep-audit:
    post:
      tags: [Deep Audit]
      summary: Run a deep code audit
      description: |
        Triggers a multi-phase deep audit of a GitHub repository:
        1. **Discovery** — Scans entire file tree via GitHub Trees API
        2. **Selection** — AI selects top N files most relevant for production readiness audit
        3. **Loading** — Fetches selected file contents (with size limits)
        4. **Analysis** — Batched LLM analysis of code quality, tests, security, completeness
        5. **Consolidation** — Merges batch results into a final comprehensive report

        Returns immediately with a running audit record. Progress is streamed via SSE.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sourceUrl]
              properties:
                sourceUrl: { type: string, format: uri }
                branch: { type: string, default: main }
                maxFiles: { type: integer, default: 200, minimum: 10, maximum: 500, description: Max files for AI to select }
                maxContentKB: { type: integer, default: 500, minimum: 100, maximum: 2000, description: Total content size limit in KB }
                provider: { type: string, enum: [claude, gemini, grok], default: claude }
                model: { type: string }
            examples:
              default: { summary: Default deep audit, value: { sourceUrl: "https://github.com/example-org/my-app" } }
              custom: { summary: Custom options, value: { sourceUrl: "https://github.com/example-org/my-app", branch: dev, maxFiles: 300, provider: claude, model: "claude-sonnet-4-20250514" } }
      responses:
        '201':
          description: Deep audit queued (running in background)
...
  /sharepoint/projects/{projectId}/audit:
    parameters:
      - { name: projectId, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [SharePoint]
      summary: Run a SharePoint deep content audit
      x-required-roles: [runner, project_lead, admin]
      description: |
        **Requires role: runner, project_lead, or admin.**

        5-phase pipeline mirroring GitHub deep audits:
        1. **Discovery** — Enumerate all SharePoint folders and files
        2. **Selection** — Prioritize files by name/type/folder (requirements, architecture, plans rank highest)
        3. **Loading** — Read content of selected files in batches (text-based files only, up to maxContentKB per file)
        4. **Analysis** — LLM analyzes content against project metadata, module structure, and velocity step completeness
        5. **Consolidation** — Save structured results (module completeness, key artifacts, recommendations)
        Returns immediately with an audit ID; processing runs asynchronously with SSE progress events.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                provider: { type: string, enum: [claude, gemini, grok], description: LLM provider for deep content analysis }
                model: { type: string, description: Specific model override }
                maxFiles: { type: integer, default: 100, description: Maximum files to read content from }
                maxContentKB: { type: integer, default: 300, description: Maximum KB to read per file }
      responses:
        '202':
          description: Audit started
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: object
                    properties:
                      auditId: { type: string, format: uuid }
                      status: { type: string, enum: [running] }
...